Add files via upload
The primary purpose of this notebook is to develop a deep learning-based system for automated breast cancer detection. It aims to achieve this by using mammogram images as input, the notebook trains models (CNN and U-Net) to accurately identify and segment the regions of interest (ROIs) that potentially contain tumors. This is achieved by predicting tumor masks, which highlight these areas.